backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Pinecone (Vector DB)
pinecone-client>=3.0.0

# LangChain (RAG framework)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3

# PDF parsing
pypdf>=4.0.0

# Google Cloud (for storage and auth)
google-auth>=2.0.0
google-cloud-storage>=2.10.0
requests>=2.28.0

# OpenAI
httpx>=0.26.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
